% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para.R
\name{checkParaCharacter}
\alias{checkParaCharacter}
\title{Check parameter value of character type whether is legal for the request of this library}
\usage{
checkParaCharacter(key, value, support = NULL)
}
\arguments{
\item{key}{The parameter name}

\item{value}{The parameter value which could be any type}

\item{support}{A vector of the supported value of parameter key}
}
\value{
True if the value is legal for parameter key
}
\description{
Check parameter value whether is legal for the request of this library
}
