% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para.R
\name{checkParaNormal}
\alias{checkParaNormal}
\title{Check parameter value of normal type whether is legal for the request of this library}
\usage{
checkParaNormal(key, value, support = NULL, prange = NULL, len = 2)
}
\arguments{
\item{key}{The parameter name}

\item{value}{The parameter value which could be any type}

\item{support}{A vector of the supported value of parameter key}

\item{prange}{The range of parameter key if value is a numeric and the length of value equal to one}

\item{len}{The legal length of parameter key if the length of value is more than one}
}
\value{
True if the value is legal for parameter key
}
\description{
Check parameter value whether is legal for the request of this library
}
