% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para.R
\name{checkParaPath}
\alias{checkParaPath}
\title{Check parameter value of path type whether is legal for the request of this library}
\usage{
checkParaPath(key, value)
}
\arguments{
\item{key}{The parameter name}

\item{value}{The parameter value which could be any type}
}
\value{
True if the value is legal for parameter key
}
\description{
Check parameter value whether is legal for the request of this library
}
