% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para.R
\name{checkParaVector}
\alias{checkParaVector}
\title{Check parameter value of vector type whether is legal for the request of this library}
\usage{
checkParaVector(key, value, len = 2)
}
\arguments{
\item{key}{The parameter name}

\item{value}{The parameter value which could be any type}

\item{len}{The legal length of parameter key if the length of value is more than one}
}
\value{
True if the value is legal for parameter key
}
\description{
Check parameter value whether is legal for the request of this library
}
