% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineObsSim.R
\name{combineObsSim}
\alias{combineObsSim}
\title{Combine the simulated and observed data}
\usage{
combineObsSim(obs, sim, variables, identifier = NULL, isOutputObsData = FALSE)
}
\arguments{
\item{obs}{A data.frame of the observed data}

\item{sim}{A data.frame of the simulated data}

\item{variables}{A list of intersting variable name}

\item{identifier}{A vector of the column names of identifier}

\item{isOutputObsData}{Whether output observed data}
}
\value{
A data.frame contian all simulated and observed data. 
The several column at the beginning are the identifier of data/
The other columns were the simulated value, the mean of observed value and the 
standard deviations of observed value for each interesting variable.
}
\description{
Calculate the mean and standard deviations of observed data.
Insert the observed mean and standard deviations into the table of simulated data
}
