% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionLibrary.R
\name{crown_temp_nwheat}
\alias{crown_temp_nwheat}
\title{Crown temperature from nwheat}
\usage{
crown_temp_nwheat(maxt, mint, snow = 0)
}
\arguments{
\item{maxt}{Maximum temperature (oC)}

\item{mint}{Minimum temperature (oC)}

\item{snow}{Snow depth (cm)}
}
\description{
Crown temperature from nwheat
}
