% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para.R
\name{defaultPara}
\alias{defaultPara}
\title{Get the all/a default parameter value}
\usage{
defaultPara(para.name = NULL)
}
\arguments{
\item{para.name}{The paramer with this name will be returned}
}
\value{
The all default parameter will be returned if the para.name is not specified, or
just the default paramter whose name is para.name
}
\description{
Get the all/a default parameter value
}
