% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionLibrary.R
\name{interpolationFunction}
\alias{interpolationFunction}
\title{Return a y value from a linear interpolation function}
\usage{
interpolationFunction(x, y, values, split = "\\\\s+")
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{values}{values}

\item{split}{split}
}
\description{
Return a y value from a linear interpolation function
}
