% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.r
\name{left}
\alias{left}
\title{Get the several characters of a string from left}
\usage{
left(str, num = 1)
}
\arguments{
\item{str}{Input string}

\item{num}{The number of character will be returned from left}
}
\value{
the character vector of the string from left
}
\description{
Get the several characters of a string from left
}
