% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.r
\name{omitBlank}
\alias{omitBlank}
\title{Omit the blank of a string}
\usage{
omitBlank(str)
}
\arguments{
\item{str}{Input string}
}
\value{
the character vector except the blank
}
\description{
Omit the blank of a string
}
