% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.r
\name{omitBlankSE}
\alias{omitBlankSE}
\title{Omit the start and end blank charcter}
\usage{
omitBlankSE(str)
}
\arguments{
\item{str}{Input string}
}
\value{
the character vector except the blank at the start and end
}
\description{
Omit the start and end blank charcter
}
