% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraphCombinations.R
\name{plotGraphCombinations}
\alias{plotGraphCombinations}
\title{Plot figures by combinations for all treatments}
\usage{
plotGraphCombinations(
  sim.obs,
  graph.combinations = defaultPara("graph.combinations"),
  identifiers = defaultPara("identifiers"),
  variables = defaultPara("variables"),
  xvar = NULL,
  isOutputObsData = FALSE
)
}
\arguments{
\item{sim.obs}{The simulated values which combine the observed values}

\item{graph.combinations}{A list to describe how to combime figures.}

\item{identifiers}{A vector of the column names of identifier.}

\item{variables}{The compared column names of observed and simulated data.}

\item{xvar}{The variable which will be drawn in x axis.}

\item{isOutputObsData}{Whether output observed data.}
}
\value{
A vector of all figures
}
\description{
Plot figures by combinations for all treatments
}
