% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraphCombinations.R
\name{plotGraphCombinationsByTreatment}
\alias{plotGraphCombinationsByTreatment}
\title{Plot figures by combinations for each treatment}
\usage{
plotGraphCombinationsByTreatment(
  pd,
  graph.combinations = NULL,
  xvar = NULL,
  xlab = NULL,
  xrange = NULL,
  yrange = NULL,
  strips = NULL,
  main = NULL,
  key = NULL,
  notes = NULL,
  isOutputObsData = FALSE,
  stage = NULL
)
}
\arguments{
\item{pd}{The simulated values which combine the observed values}

\item{graph.combinations}{A list to describe how to combime figures}

\item{xvar}{The variable which will be drawn in x axis}

\item{xlab}{the label of x axis}

\item{xrange}{The range of x variable.}

\item{yrange}{The range of y variable.}

\item{strips}{A vector of strips.}

\item{main}{The main title of the figure.}

\item{key}{A list to describe key.}

\item{notes}{A string which put in the footnoot of pages.}

\item{isOutputObsData}{Whether output observed data.}

\item{stage}{A vector to describe the stage}
}
\value{
A object of this figure
}
\description{
Plot figures by combinations for each treatment
}
