% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObsSimCompare.R
\name{plotObsSimCompare}
\alias{plotObsSimCompare}
\title{Plot the figure to compare the observed and simulated variables}
\usage{
plotObsSimCompare(sim.obs, identifiers, variables)
}
\arguments{
\item{sim.obs}{The simulated values which combine the observed values
for each interesting variable.}

\item{identifiers}{A vector of the column names of identifier}

\item{variables}{A vector of the column names of interesting variables}
}
\value{
The object of this figure
}
\description{
Plot the figure to compare the observed and simulated variables
}
