% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RApsim.R
\name{RApsim}
\alias{RApsim}
\title{This the main function to compare the simulated and observed data.}
\usage{
RApsim(
  para = NULL,
  obsFile = NULL,
  simFiles = NULL,
  simFolders = NULL,
  simSuffix = "outputfile_daily",
  xvar = "DaysAfterSowing",
  ...
)
}
\arguments{
\item{para}{The parameter list. See defaultPara for more detailed information.}

\item{obsFile}{The absolute or relative file path to the observed data}

\item{simFiles}{A vector of files of simulated results of APSIM}

\item{simFolders}{A vector of folders of simulated results of APSIM}

\item{simSuffix}{The suffix of files of simulated results of APSIM}

\item{xvar}{The variable will be put into the x-axis}

\item{...}{The parameters of setpara functions}
}
\description{
This package was originally designed to compare the observed data
and simulated results of Apsim. But you could use this package to 
compare any observed and simulated data. All figures created by 
RApsim are flexible. You can design them through several parameters.
only columns, which were indicated in "identifiers" and "variables", 
would be used. Any other data would been ignored. "graph.combinations" 
would be used to design the graph which compare simulated and observed data.
You can used "xvar" to set which data from "identifiers" was drawn in the 
x-axis.
}
\examples{
# Set met files and/or met folders
\dontrun{ RApsim( simFolders = "path-to-simfolder", obsFile = "path-to-obsfiles" ) }
# Set a specific parameter
\dontrun{ RApsim( simFolders = "path-to-simfolder", obsFile = "path-to-obsfiles", outputPrefix = "output" ) }
# Using thermal time which was drawn in x-axis
\dontrun{ RApsim( simFolders = "path-to-simfolder", obsFile = "path-to-obsfiles", xvar = "cumTT" ) }
}
\seealso{
See \code{\link{setpara}} for avaiable parameters of RApsim. 
\code{\link{readobs}} for the file format of observed files.
\code{\link{readsim}} for the file format of simulated files.
}
