% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readobs}
\alias{readobs}
\title{Read observed data from file}
\usage{
readobs(obsFile = NULL)
}
\arguments{
\item{obsFile}{The absolute or relative file path to the observed data}
}
\value{
A data frame of observed data
}
\description{
The codes will skip first two lines of the observed data file.
The third line of this file should be the name of each column.
The fourth line of this file should be the unit of each column.
The observed data must be started from the fifth line. 
The missing data should be indicated by * not "0".
}
