% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{readPara}
\alias{readPara}
\title{read a parameter from sim or module file}
\usage{
readPara(name, ...)
}
\arguments{
\item{name}{The parameter name}

\item{...}{Other arguments to specify sim file and module file}
}
\description{
read a parameter from sim or module file
}
