% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readsim}
\alias{readsim}
\title{Read simulated data of APSIM from file}
\usage{
readsim(simFiles = NULL, simFolders = NULL, simSuffix = "outputfile_daily")
}
\arguments{
\item{simFiles}{A vector of files of simulated results of APSIM}

\item{simFolders}{A vector of folders of simulated results of APSIM}

\item{simSuffix}{The suffix of files of simulated results of APSIM}
}
\value{
A data frame of simulated data
}
\description{
The head of the simulated could contain some comment information (less than 100 lines).
The comment line must format as "variable = value". The codes read all lines, 
then add to simulated data.
The first line after comments should be the name of each column.
The second line after comments should be the unit of each column.
The simulated data must be started from the third line after comments. 
If simFolders was specified, the file name of data must be like this 
"file_name-suffix.out"
}
