% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraphCombinations.R
\name{rescale}
\alias{rescale}
\title{Rescale the x to ( 0, 1 )}
\usage{
rescale(x, pars = scales.pars(x))
}
\arguments{
\item{x}{A numeric vector to rescale}

\item{pars}{The minimum values and iterated differences}
}
\value{
The new vector which all value in ( 0, 1 )
}
\description{
Rescale the x to ( 0, 1 )
}
