% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.r
\name{right}
\alias{right}
\title{Get the several characters of a string from right}
\usage{
right(str, num = 1)
}
\arguments{
\item{str}{Input string}

\item{num}{The number of character will be returned from right}
}
\value{
the character vector of the string from right
}
\description{
Get the several characters of a string from right
}
