% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraphCombinations.R
\name{scales.pars}
\alias{scales.pars}
\title{Calculate the minimum values and iterated differences}
\usage{
scales.pars(x)
}
\arguments{
\item{x}{A numeric vector or matrix containing the values to be differenced.}
}
\value{
A list contain the minimum values and iterated differences
}
\description{
Calculate the minimum values and iterated differences
}
