% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.r
\name{searchChar}
\alias{searchChar}
\title{Search a character from a string}
\usage{
searchChar(str, search, start = 1)
}
\arguments{
\item{str}{Input string}

\item{search}{The character will be searched}

\item{start}{The start position to search}
}
\value{
The postion where character firstly appear
}
\description{
Search a character from a string
}
