% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para.R
\name{setpara}
\alias{setpara}
\title{Set/Change some parameters}
\usage{
setpara(
  p = NULL,
  outputFormat = defaultPara("outputFormat"),
  outputPrefix = defaultPara("outputPrefix"),
  isOutputObsData = defaultPara("isOutputObsData"),
  identifiers = defaultPara("identifiers"),
  variables = defaultPara("variables"),
  graph.combinations = defaultPara("graph.combinations")
)
}
\arguments{
\item{p}{The old parameter list. The default parameter would be used
if p was not specified.}

\item{outputFormat}{The format of output files. Currently only pdf format is supported}

\item{outputPrefix}{The prefix of output files. A new folder will be created if the last character}

\item{isOutputObsData}{Whether output observed data
is "/" or "\"}

\item{identifiers}{A list of the column names to identify each data.
This parameter must have 3 elements: "sim, label, unit", and the three
element must have the same length. The unit element could be expression
for the complicated unit of variables. The sim should be from the simulated
files and the observed file should contain the same columns.}

\item{variables}{the file path to a csv format file which lists 
the compared column names of observed and simulated data.
This file must have 6 column: "sim, obs, label, unit, col, pch" 
and the six column must have the same length. The unit column could be 
expression for the complicated unit of variables.}

\item{graph.combinations}{A list to describe the graph which compare
the simulated and observed data.}
}
\description{
The validity of some parameters have been checked. only columns, which 
were indicated in "identifiers" and "variables", would be used. Any other
data would been ignored. "graph.combinations" would be used to design the 
graph which compare simulated and observed data.
}
