% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionLibrary.R
\name{sign_apsim}
\alias{sign_apsim}
\title{Transfer of sign - from FORTRAN.
The result is of the same type and kind as a. Its value is the abs(a) of a,
if b is greater than or equal positive zero; and -abs(a), if b is less than
or equal to negative zero.
Example a = sign_apsim (30,-2) ! a is assigned the value -30}
\usage{
sign_apsim(a, b)
}
\arguments{
\item{a}{value 1}

\item{b}{value 2}
}
\description{
Transfer of sign - from FORTRAN.
The result is of the same type and kind as a. Its value is the abs(a) of a,
if b is greater than or equal positive zero; and -abs(a), if b is less than
or equal to negative zero.
Example a = sign_apsim (30,-2) ! a is assigned the value -30
}
