% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2tem.R
\name{sim2tem}
\alias{sim2tem}
\title{Generate the template file for REMS from APSIM simulation file}
\usage{
sim2tem(sim, com, output = "output.tem")
}
\arguments{
\item{sim}{The file name of apsim}

\item{com}{The file name of commands}

\item{output}{The output file name}
}
\description{
This function is used to convert the apsim file to REMS template
file according to the description of a command file. The command
file is the csv format. In the command file, first column is the 
line number which need to deal with; the second column is the old
value in the apsim file; the third column is the new value in the 
REMS template file. If the old and new value are empty, this row
will be deleted from output file. If the old value is empty, but 
the new value is not empty, a new line will be added before this
line. If the old and new value are not empty, the old value will
be replaced by the new value.
}
\examples{
\dontrun{ sim2tem("test.apsim","test.csv") }
}
