% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.r
\name{splitEqual}
\alias{splitEqual}
\title{Split a string contained a equal mark}
\usage{
splitEqual(vector)
}
\arguments{
\item{vector}{The input vector to convert}
}
\value{
the variable name and value
}
\description{
Split a string contained a equal mark
}
