% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionLibrary.R
\name{sunRiseSet}
\alias{sunRiseSet}
\title{The time for sunrise and sunset
from 90 degree in am and pm. +ve above the horizon, -ve below the horizon.}
\usage{
sunRiseSet(doy, lat, long, timezone = 0, angle = 6, flag = "rise")
}
\arguments{
\item{doy}{day of year number}

\item{lat}{latitude of site (deg)}

\item{long}{longitude of site (deg)}

\item{timezone}{local time}

\item{angle}{angle to measure time between, such as twilight (deg).
angular distance between 90 deg and end of twilight - altitude of sun. +ve up, -ve down.}

\item{flag}{rise or set}
}
\description{
The time for sunrise and sunset
from 90 degree in am and pm. +ve above the horizon, -ve below the horizon.
}
