% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.r
\name{vector2string}
\alias{vector2string}
\title{Convert vector to string}
\usage{
vector2string(vector, sep = ", ")
}
\arguments{
\item{vector}{The input vector to convert}

\item{sep}{The separater with default value ", "}
}
\value{
The string which connected all members of vector
}
\description{
Convert vertor to string with a separater
}
