% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionLibrary.R
\name{wheat_de_vernaliz_days}
\alias{wheat_de_vernaliz_days}
\title{Calculate daily vernalisation and accumulate to g_cumvd}
\usage{
wheat_de_vernaliz_days(maxt, mint, tempcr, g_cumvd, dlt_cumvd)
}
\arguments{
\item{maxt}{Daily maximum Temperature}

\item{mint}{Daily minimum temperature}

\item{tempcr}{Crown temperature}

\item{g_cumvd}{cumulative vernalisation days till yesterday}

\item{dlt_cumvd}{delta cummulative vernalisation days}
}
\description{
Calculate daily vernalisation and accumulate to g_cumvd
}
