% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionLibrary.R
\name{wheat_photoperiod_effect}
\alias{wheat_photoperiod_effect}
\title{Photoperiod factor}
\usage{
wheat_photoperiod_effect(photoperiod, p_photop_sen)
}
\arguments{
\item{photoperiod}{photoperiod}

\item{p_photop_sen}{p_photop_sen}
}
\description{
Photoperiod factor
}
