% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionLibrary.R
\name{wheat_vernaliz_days}
\alias{wheat_vernaliz_days}
\title{Calculate daily vernalisation and accumulate to g_cumvd}
\usage{
wheat_vernaliz_days(maxt, mint, tempcr)
}
\arguments{
\item{maxt}{Daily maximum Temperature}

\item{mint}{Daily minimum temperature}

\item{tempcr}{Crown temperature}
}
\description{
Calculate daily vernalisation and accumulate to g_cumvd
}
