% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionLibrary.R
\name{wheat_vernaliz_effect}
\alias{wheat_vernaliz_effect}
\title{Vernalisation factor}
\usage{
wheat_vernaliz_effect(p_vern_sens, cumvd, dlt_cumvd, reqvd)
}
\arguments{
\item{p_vern_sens}{p_vern_sens}

\item{cumvd}{cumvd}

\item{dlt_cumvd}{dlt_cumvd}

\item{reqvd}{reqvd}
}
\description{
Vernalisation factor
}
