% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheatPhenology.R
\name{wheatPhenology}
\alias{wheatPhenology}
\title{APSIM-Wheat module}
\usage{
wheatPhenology(
  sim,
  stages = 6,
  stageNames = "flow",
  model = "APSIM",
  met = NULL,
  daily = FALSE,
  ...
)
}
\arguments{
\item{sim}{The path to sim file.}

\item{stages}{The stages of outputs}

\item{stageNames}{The stage names of outputs}

\item{model}{The model of Phenology. APSIM for default method in APSIM,
V1 and V2 change the minimum function to multiple. V2 extend photoperiod
effect of flowering}

\item{met}{The weather data with weaana class.}

\item{daily}{whether output daily stages}

\item{...}{Other arguments}
}
\description{
APSIM-Wheat module
}
\examples{
met <- system.file("extdata/weather.met", package = "APSIMWheatPhenology")
sim <- system.file("extdata/example.sim", package = "APSIMWheatPhenology")
# Flowering time using parameter in the sim file
df <- wheatPhenology(sim, stages = 6, stageNames = 'flow', model = 'APSIM',
                     met = met, daily = FALSE)
head(df)

# Models V1 or V2 can be used
df <- wheatPhenology(sim, stages = 6, stageNames = 'flow', model = 'V1',
                     met = met, daily = FALSE)
head(df)

# Weather can be passed as a weaana object
met_data <- weaana::readWeatherRecords(met)
df <- wheatPhenology(sim, stages = 6, stageNames = 'flow', model = 'APSIM',
                     met = met_data, daily = FALSE)
head(df)

# Multiple stages can be calculated
df <- wheatPhenology(sim, stages = c(5.74, 6), stageNames = c("head", 'flow'),
                     model = 'APSIM',
                     met = met_data, daily = FALSE)
head(df)

# Daily output can be exported
df <- wheatPhenology(sim, stages = c(5.74, 6), stageNames = c("head", 'flow'),
                     model = 'APSIM',
                     met = met_data, daily = TRUE)
head(df)

# Multiple simulations can be specified in a factor
# Any parameters related with plant can be used
factors <- expand.grid(list(vern_sens = seq(1, 5), photop_sens = seq(1, 5)))
df <- wheatPhenology(sim, stages = c(5.74, 6), stageNames = c("head", 'flow'),
                     model = 'APSIM',
                     met = met_data, daily = FALSE,
                     factors = factors)
head(df)
}
