% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheatPhenology.R
\name{wheatPhenology}
\alias{wheatPhenology}
\title{APSIM-Wheat module}
\usage{
wheatPhenology(
  sim,
  stages = 6,
  stageNames = "flow",
  model = "APSIM",
  met = NULL,
  daily = FALSE,
  ...
)
}
\arguments{
\item{sim}{The path to sim file.}

\item{stages}{The stages of outputs}

\item{stageNames}{The stage names of outputs}

\item{model}{The model of Phenology. APSIM for default method in APSIM,
V1 and V2 change the minimum function to multiple. V2 extend photoperiod
effect of flowering}

\item{met}{The weather data with weaana class.}

\item{daily}{whether output daily stages}

\item{...}{Other arguments}
}
\description{
APSIM-Wheat module
}
